Refactor: input과 등록 버튼을 TextBox컴포넌트로 분리
